Submission for Assignment 05

+++ test/tests.sh
@@ -12,16 +12,49 @@ for file in ../inputs/*.in; do
   noextname=${name%.*}
   echo $noextname
   cd lexer-outputs
+
+  echo "Testing Lexing"
   eval stack exec aardvark-exe \-\- -l -r $file > temp.test
-  diff temp.test $noextname.lex
+  diff temp.test $noextname.lex > diff.output
+  size=$(wc -c <"diff.output")
+  if [ $size -gt 0 ]; then
+      echo "Failed test."
+      echo "Expected:"
+      cat $noextname.lex
+      echo "Actual:"
+      cat temp.test
+      exit 2
+  fi
   rm temp.test
+
+  echo "Testing parsing"
   cd ../parser-outputs
   eval stack exec aardvark-exe \-\- -p -r $file > temp.test
-  diff temp.test $noextname.parsed
+  diff temp.test $noextname.parsed > diff.output
+  size=$(wc -c <"diff.output")
+  if [ $size -gt 0 ]; then
+      echo "Failed test."
+      echo "Expected:"
+      cat $noextname.parsed
+      echo "Actual:"
+      cat temp.test
+      exit 2
+  fi
   rm temp.test
+
+  echo "Testing evaluation"
   cd ../eval-outputs
   eval stack exec aardvark-exe \-\- -r $file > temp.test
-  diff temp.test $noextname.evaluated
+  diff temp.test $noextname.evaluated > diff.output
+  size=$(wc -c <"diff.output")
+  if [ $size -gt 0 ]; then
+      echo "Failed test."
+      echo "Expected:"
+      cat $noextname.evaluated
+      echo "Actual:"
+      cat temp.test
+      exit 2
+  fi
   rm temp.test
 done
